requirements.txt
python-telegram-bot==3.2.0
python-quran-odoa==1.3
